implement provider strategy continue refactor
